+++ src/fetch/types.ts
@@ -6,7 +6,7 @@ import type {
 	IsNever,
 	Prettify,
 	TreatyToPath,
-	ThrowHttpErrors
+	ThrowHttpError
 } from '../types'
 
 export namespace EdenFetch {
@@ -22,7 +22,7 @@ export namespace EdenFetch {
 
 	export interface Config extends RequestInit {
 		fetcher?: typeof globalThis.fetch
-		throwHttpErrors?: ThrowHttpErrors
+		throwHttpError?: ThrowHttpError
 	}
 
 	export type Fn<Schema extends Record<string, any>> = <
@@ -63,7 +63,7 @@ export namespace EdenFetch {
 			(IsUnknown<Route['body']> extends false
 				? { body: Route['body'] }
 				: { body?: unknown }) & {
-          throwHttpErrors?: ThrowHttpErrors
+          throwHttpError?: ThrowHttpError
         }
 	) => Promise<
 		Prettify<